Add a delay after clearing the column pin.
After 4e3f7d7 and a9b12fa, it seems the pin code is executing faster, so
exposed a bug that PSOR and PCOR don't update the pin instantly.

For now it's a 1 ms delay. This should be updated to a smaller delay, or
figure out if there is a way to check if the pin finished updating.